Feature DPoS - Extract most of logics of dpos from contracts and ready to refactor consensus part
