+++ AElf.Kernel.Types/Globals.cs
@@ -7,5 +7,6 @@
         public static readonly string SmartContractZeroIdString = "__SmartContractZero__";
         public static readonly int MiningTime = 8000;
         public static readonly int CheckTime = 2000;
+        public static readonly int WaitFirstRoundTime = 6000;
     }
 }
