+++ AElf.Node/Node/MainChainNode.cs
@@ -55,8 +55,8 @@ namespace AElf.Kernel.Node
 
         private readonly IBlockExecutor _blockExecutor;
 
-        private DPoS _dPoS;
-        private DPoSCheck _dPoSCheck;
+        private DPoSTxGenerator _dPoSTxGenerator;
+        private DPoSHelper _dPoSHelper;
 
         public Hash ContractAccountHash =>
             _chainCreationService.GenesisContractHash(_nodeConfig.ChainId);
@@ -713,8 +713,8 @@ namespace AElf.Kernel.Node
             {
                 _logger?.Debug("-- DPoS Mining Has been fired!");
                 
-                _dPoS = new DPoS(_nodeKeyPair);
-                _dPoSCheck = new DPoSCheck(_worldStateDictator, _nodeKeyPair, ChainId, BlockProducers, ContractAccountHash, _logger);
+                _dPoSTxGenerator = new DPoSTxGenerator(_nodeKeyPair);
+                _dPoSHelper = new DPoSHelper(_worldStateDictator, _nodeKeyPair, ChainId, BlockProducers, ContractAccountHash, _logger);
 
                 //Record the rounds count in local memory
                 ulong roundsCount = 0;
@@ -750,7 +750,7 @@ namespace AElf.Kernel.Node
                             return;
                         }
 
-                        var actualRoundsCount = x == 0 ? 0 : await GetActualRoundsCount();
+                        var actualRoundsCount = x == 0 ? 0 : _dPoSHelper.RoundsCount.Value;
                         if (roundsCount != actualRoundsCount)
                         {
                             //Update the rounds count
@@ -770,7 +770,7 @@ namespace AElf.Kernel.Node
                                 return;
                             }
                             
-                            var dpoSInfo = await ExecuteTxsForFirstExtraBlock();
+                            var dpoSInfo = ExecuteTxsForFirstExtraBlock();
 
                             await BroadcastSyncTxForFirstExtraBlock(dpoSInfo);
 
@@ -855,7 +855,7 @@ namespace AElf.Kernel.Node
                             if (lastTryToPublishInValueRoundsCount != roundsCount)
                             {
                                 //Try to publish in value (every BP can do this)
-                                await BroadcastTransaction(_dPoS.GetTxToPublishInValueTx(
+                                await BroadcastTransaction(_dPoSTxGenerator.GetTxToPublishInValueTx(
                                     await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash, inValue,
                                     new UInt64Value {Value = roundsCount}));
 
@@ -951,40 +951,12 @@ namespace AElf.Kernel.Node
                 
             });
         }
-       
-        private async Task<DPoSInfo> ExecuteTxsForFirstExtraBlock()
-        {
-            var txsForFirstExtraBlock = _dPoS.GetTxsForFirstExtraBlock(
-                await GetIncrementId(_nodeKeyPair.GetAddress()), BlockProducers, ContractAccountHash
-            );
 
-            var dPoSInfo = new DPoSInfo();
-            
-            foreach (var tx in txsForFirstExtraBlock)
-            {
-                var tc = new TransactionContext
-                {
-                    Transaction = tx
-                };
-                Executive.SetTransactionContext(tc).Apply(true).Wait();
-                
-                if (tx.MethodName.StartsWith("Generate"))
-                {
-                    if (!tc.Trace.IsSuccessful())
-                    {
-                        _logger?.Debug(tc.Trace.StdErr);
-                    }
-
-                    dPoSInfo = tc.Trace.RetVal.Data.DeserializeToPbMessage<DPoSInfo>();
-                }
-            }
-
-            return dPoSInfo;
-        }
+        #region Broadcast Txs
 
         private async Task BroadcastSyncTxForFirstExtraBlock(DPoSInfo dPoSInfo)
         {
-            var txToSyncFirstExtraBlock = _dPoS.GetTxToSyncFirstExtraBlock(
+            var txToSyncFirstExtraBlock = _dPoSTxGenerator.GetTxToSyncFirstExtraBlock(
                 await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash, dPoSInfo, BlockProducers);
 
             await BroadcastTransaction(txToSyncFirstExtraBlock);
@@ -992,7 +964,7 @@ namespace AElf.Kernel.Node
 
         private async Task BroadcastTxsForNormalBlock(ulong roundsCount, Hash outValue, Hash signature, ulong incrementId)
         {
-            var txForNormalBlock = _dPoS.GetTxsForNormalBlock(
+            var txForNormalBlock = _dPoSTxGenerator.GetTxsForNormalBlock(
                 incrementId, ContractAccountHash, roundsCount,
                 outValue, signature);
             foreach (var tx in txForNormalBlock)
@@ -1001,76 +973,31 @@ namespace AElf.Kernel.Node
             }
         }
         
-        private async Task<Tuple<RoundInfo, RoundInfo, StringValue>> ExecuteTxsForExtraBlock(ulong incrementId)
-        {
-            var txsForNormalBlock = _dPoS.GetTxsForExtraBlock(
-                incrementId, ContractAccountHash);
-            
-            var currentRoundInfo = new RoundInfo();
-            var nextRoundInfo = new RoundInfo();
-            // ReSharper disable once InconsistentNaming
-            var nextEBP = new StringValue();
-            
-            foreach (var tx in txsForNormalBlock)
-            {
-                var tc = new TransactionContext
-                {
-                    Transaction = tx
-                };
-                Executive.SetTransactionContext(tc).Apply(true).Wait();
-
-                if (!tc.Trace.StdErr.IsNullOrEmpty())
-                {
-                    _logger?.Error(tc.Trace.StdErr);
-                    continue;
-                }
-                
-                if (tx.MethodName.StartsWith("Supply"))
-                {
-                    currentRoundInfo = tc.Trace.RetVal.Data.DeserializeToPbMessage<RoundInfo>();
-                }
-                
-                if (tx.MethodName.StartsWith("Generate"))
-                {
-                    nextRoundInfo =tc.Trace.RetVal.Data.DeserializeToPbMessage<RoundInfo>();
-                }
-
-                if (tx.MethodName.StartsWith("Set"))
-                {
-                    nextEBP = tc.Trace.RetVal.Data.DeserializeToPbMessage<StringValue>();
-                }
-            }
-
-            return Tuple.Create(currentRoundInfo, nextRoundInfo, nextEBP);
-        }
-
         // ReSharper disable once InconsistentNaming
         private async Task BroadcastTxsToSyncExtraBlock(ulong incrementId,
             RoundInfo currentRoundInfo, RoundInfo nextRoundInfo, StringValue nextEBP)
         {
-            var txForExtraBlock = _dPoS.GetTxToSyncExtraBlock(
+            var txForExtraBlock = _dPoSTxGenerator.GetTxToSyncExtraBlock(
                 incrementId, ContractAccountHash, currentRoundInfo, nextRoundInfo, nextEBP);
 
             await BroadcastTransaction(txForExtraBlock);
         }
 
-        private async Task<ulong> GetActualRoundsCount()
+        #endregion
+        
+        private DPoSInfo ExecuteTxsForFirstExtraBlock()
         {
-            var tcGetRoundsCountTx = new TransactionContext
-            {
-                Transaction = _dPoS.GetRoundsCountTx(await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash)
-            };
-            Executive.SetTransactionContext(tcGetRoundsCountTx).Apply(true).Wait();
+            return _dPoSHelper.GenerateInfoForFirstTwoRounds();
+        }
+        
+        private async Task<Tuple<RoundInfo, RoundInfo, StringValue>> ExecuteTxsForExtraBlock(ulong incrementId)
+        {
+            var currentRoundInfo = await _dPoSHelper.SupplyPreviousRoundInfo();
+            var nextRoundInfo = await _dPoSHelper.GenerateNextRoundOrder();
+            // ReSharper disable once InconsistentNaming
+            var nextEBP = await _dPoSHelper.SetNextExtraBlockProducer();
             
-            if (tcGetRoundsCountTx.Trace.IsSuccessful())
-            {
-                return tcGetRoundsCountTx.Trace.RetVal.Data.DeserializeToUInt64();
-            }
-            
-            _logger?.Debug("Failed to get rounds count");
-            _logger?.Error(tcGetRoundsCountTx.Trace.StdErr);
-
-            return 0;
+            return Tuple.Create(currentRoundInfo, nextRoundInfo, nextEBP);
         }
 
         // ReSharper disable once InconsistentNaming
@@ -1079,7 +1006,7 @@ namespace AElf.Kernel.Node
             // ReSharper disable once InconsistentNaming
             var tcGetDPoSInfo = new TransactionContext
             {
-                Transaction = _dPoS.GetDPoSInfoToStringTx(await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash)
+                Transaction = _dPoSTxGenerator.GetDPoSInfoToStringTx(await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash)
             };
             Executive.SetTransactionContext(tcGetDPoSInfo).Apply(true).Wait();
 
@@ -1103,30 +1030,14 @@ namespace AElf.Kernel.Node
 
         private async Task<Hash> CalculateSignature(Hash inValue)
         {
-            try
-            {
-                var tcCalculateSignature = new TransactionContext
-                {
-                    Transaction =
-                        _dPoS.GetCalculateSignatureTx(
-                            await GetIncrementId(_nodeKeyPair.GetAddress()), ContractAccountHash,
-                            inValue)
-                };
-                Executive.SetTransactionContext(tcCalculateSignature).Apply(true).Wait();
-                return tcCalculateSignature.Trace.RetVal.Data.DeserializeToPbMessage<Hash>();
-            }
-            catch (Exception e)
-            {
-                _logger?.Error(e, "Failed to calculate signature");
-                return Hash.Generate();
-            }
+            return await _dPoSHelper.CalculateSignature(inValue);
         }
 
         private async Task<bool> CheckAbleToMineNormalBlock()
         {
             try
             {
-                return await _dPoSCheck.AbleToMine();
+                return await _dPoSHelper.AbleToMine();
             }
             catch (Exception e)
             {
@@ -1139,7 +1050,7 @@ namespace AElf.Kernel.Node
         {
             try
             {
-                return await _dPoSCheck.ReadyForHelpingProducingExtraBlock();
+                return await _dPoSHelper.ReadyForHelpingProducingExtraBlock();
             }
             catch (Exception e)
             {
@@ -1152,7 +1063,7 @@ namespace AElf.Kernel.Node
         {
             try
             {
-                return await _dPoSCheck.IsTimeToProduceExtraBlock();
+                return await _dPoSHelper.IsTimeToProduceExtraBlock();
             }
             catch (Exception e)
             {
@@ -1165,7 +1076,7 @@ namespace AElf.Kernel.Node
         {
             try
             {
-                return await _dPoSCheck.AbleToProduceExtraBlock();
+                return await _dPoSHelper.AbleToProduceExtraBlock();
             }
             catch (Exception e)
             {
